Update theme views to prompt user for overwriting static files
Loading a theme would normally copy images and scripts to the /media/
dirs.  In many cases users have customized those files (e.g. using the
admin filebrowser) and don't want their changes to be overwritten.  The
typical case is a chapter's logo image.

The solution presented here is to create a new view, confirm_overwrite().
This view is called whenever a difference is detected between the theme
provided files and the files currently in /media.  In this case, the user
has to choose whether to keep or overwrite each file (the default is
keep).  With this form submitted, confirm_overwrite hands control back
to the previous view function (e.g. select() or recompile()) which tells
the ThemeController which files to keep.