Always use compact_classes display
The class change grid is not a replacement for the catalog.  It is not
designed to show complete descriptions of all classes in a searchable
fashion.  Thanks to Gary's changes in 4395f56, you can show the class
titles if you want... but I think it's a little excessive to have the
entire description there.

This also makes it possible to revert the grid to its initial display
mode.