Fix test failure in Ajax scheduling: testTooOldChangeLog
The test code was not actually doing anything after clearing the change log,
yet it expected that the view would request a reload.  In fact the view code
is smarter than that and only requests a reload if something has changed
since you deleted the change log.  So we have to test that behavior.