Remove esp.twilltests app
The code is 5+ years old and has not been used recently.  There are
now other tools for testing against a live server, which should not
be in our test suite, and we will have to come up with a better
solution for testing against a dev server running "in parallel"
with the test suite (like our Selenium tests).