Excludes reversion unit tests from running
The "RevisionMiddleware can only be included in MIDDLEWARE_CLASSES
once" failure was due to a faulty assumption in the Reversion
unit tests. The test view added the RevisionMiddleware as a
decorator, so because our app also addes the RevisionMiddleware
to MIDDLEWARE_CLASSES, we end up using it twice. No simple
workaround comes to mind, so I'm disabling all reversion tests.