Makes QSD does not exist 404 properly.
In the past, the "this page does not exist, would you like to create
it?" page for admins was returning as 200, meaning that if there was a
redirect away from a (nonexistent) QSD page, but you had the ability to
create it, the redirect would be ignored.  Now it will 404, still
returning the same message, so that the redirect middleware can catch
the 404 and return the redirect.