Changes dashboard edit button to proper link
Using Javascript for the link was unnecessary and prevented opening the
edit page in a new tab.

Changes the delete button back to an input button. Github issue #244
was incorrect to suggest that it be changed to a link, since it is an
action.

Revert "Fix for issue #244 - Change class delete, edit buttons to links"

This reverts commit a147257c0b0fa6e58690a274c4fa2c66d7ebf8f2.