Fixes Markdown issues in default inline QSD
Fixes unintended usage of Markdown code blocks in default inline QSD

Because of the way templates are usually written, there will often be
unintended whitespace at the beginning of the default content of an
inline QSD. Usually the block start tag will be followed by a linebreak,
some indentation, and then the actual content. However, Markdown will
interpret this as a code block.  To avoid this, we assume that the
default content will never purposely start with a Markdown code block,
and we strip this unintended space.

Also removes some whitespace from studentregcore/mainpage.html.

Resolves Github issue #997.