Implements a @secure_required decorator
This decorator returns a view that can only be accessed via a secure
request. If the request is not secure, the wrapped view redirects to the
https version of the uri.

Replaces financialaidappmodule/aid_direct.html with this decorator, and
removes the finaid_directions_step Tag.

Partially resolves Github issue #1007.