Gives admins implicit permissions
Since the admins field is no longer part of program creation, this makes
it easier for admins to be able to have deadline permissions on a
program: they will always have them. In the old schema, being an
administrator was the same as having recursive permissions, so this
change is a revert back to old behavior.