Some minor adjustments to the text messaging module.
The user selection now uses the regular user search controller instead of selecting specific lists. Instead of printing to console, what was printed will instead be displayed on the finalization page. The user interface is now three-stage instead of two-stage, with an intermediate stage where the user can double-check the number of users who will be sent a text message.