Several fixes to RegistrationType handling in Two-Phase Student Reg
1. Use reg_type.__unicode__() rather than reg_type.displayName
directly (in case .displayName is not set, we should default to .name,
which .__unicode__() already does).

2. Use RegistrationType.get_or_create in /rank_classes view, so that
the page doesn't crash if one of the expected RegistrationTypes
doesn't exist yet.

3. Don't use get_or_create in /save_priorities view. Previously this
was a minor security hole that allowed students to e.g. register for a
class with a RegistrationType of Priority/{arbitrary injected string}
by submitting tampered post data.

4. Fix a jQuery selector so that the page doesn't try to submit empty
registrations with Priority/undefined.