Fixes QSD saving rendered versions of HTML special characters
Previously, the direct QSD editing textarea used the non-escaped
version of the content to put as the starting point for the textarea,
which resulted in special characters being saved as their rendered
form on the second save of a QSD page. This change makes those
contents escaped. The static edit page was already doing this
properly.

See #974 for more details.