Update make_virtualenv.sh to support different directory locations
On Vagrant, you can't create a virtualenv within the working copy (at
least within the current design, which uses a shared folder for
convenience).  So, allow make_virtualenv.sh to accept the virtualenv path
as an argument (if there is no argument, it uses the normal location).
Also allow a --virtualenv option to be supplied to update_deps.sh which
passes it along.