Add a function to check setup of credit card module.
This is mainly for admins, who may add the credit card module to a program
and not know that any more setup needs to be done.

More setup does need to be done, since we can't publicly distribute Stripe
account keys.  However, we can put keys in local_settings.py on the
production servers, in which case admins won't need to think about it.