Remove errant first_slot parameter from PDF catalog renderer
The printables module has a catalog view which lets you select the
ordering of classes and then generate a PDF catalog.  The options
first_sort, second_sort and third_sort are used by this view.  When
rendering the catalog, these options are provided to the coursecatalog
view as a GET parameter sort_name_list.

In f89d695 I mistakenly thought that first_sort was provided as a
parameter to the coursecatalog view (the actual Latex renderer), when in
fact it isn't.  This commit fixes that by relying entirely on
sort_name_list.