Releases the dbmail lock after use
In theory this should be handled automatically after the process ends,
but for some reason I encountered a lock that did not behave this way
during testing.

I may or may not have been doing something wrong. But either way, there
is no harm in manually unlocking when done. Better safe than sorry.