Fixes multi-day teacher check-in
Only checkins from the date of 'when' (which defaults to the current
date) are considered, so that check-in can be reused each day/week of
the program. To see who failed to check in on a certain day / by a
certain time, admins can use the pre-existing "previous day" feature.

Resolves Github issue #818.