Remove unused dblog code.
Looks like this app was designed to log errors in the database.  The last time
one was logged on esp.mit.edu was in 2010, so I don't think we're actually
using this functionality; let's kill it!

It was called in a few places:
* users.getTaughtClassesFromProgram attempted to log, but not raise, an error.
  This is really bad anyway so it now raises ESPError.
* the ESPError middleware attempted to save the error to the database after
  sending it to serverlog, but it was wrapped in a try-except block, and
  doesn't seem to have been happening.
* Some other navbar view imported, but did not use it.  That entire view seems
  to be unused, so I deleted it too.

I think in the long term we should probably improve our logging setup, possibly
involving a new-and-improved way of logging errors to the DB, but I don't think
we will actually want to build on this.

Solves #979.