Reconfigure Stripe module to keep donation amount synced with server
Ben pointed out that, previously, the Stripe module made it possible for
the client to "choose their amount" by manipulating POST variables.  That
was because our form was handling the donation checkbox client-side, and
the server did not know how much the user chose to donate (only our Stripe
records would tell us that).

This commit fixes the problem by rewriting much of the module so that any
changes to donation preferences are immediately sent to the
set_donation_amount JSON view, which creates the necessary Transfer[s].
Then before charging the credit card, the amount submitted via POST
(through Stripe) is checked against the amount the accounting system
thinks the user owes.  This should prevent any inconsistencies created by
clever students.

Also updates the "pitch" for LU to what Ben suggested, though this may
change in the future.