Fix how JSONDataModule is added to created program
Refactor how the JSONDataModule is added to the program_modules field of
the ProgramCreationForm.  Perform the addition in a clean_FIELD()
method, rather than directly in setup.py.

By default, a Django ModelForm will take the raw input for a
MultipleChoiceField and coerce it into a list of strings, even if the
real data should be something else (in this case, a list of integers).
The original code checked this list for the id (an integer) of the
JSONDataModule, and added the module if it was missing from the list.

In theory, this should lead to an error.  The integer would always be
missing from the list of strings, so if the JSONDataModule was already
in the list (which is always the case during testing, or when the admin
creating the program correctly selects it), it would get added a second
time.  Then, when the program_modules field gets saved into the new
Program object, an error would be generated for having a duplicate entry
in the many-to-many relation.

For some reason that I don't know, this error never actually manifested
itself previously.  However, I ran into it when making changes to how
testing runs.

This refactoring fixes the problem I ran into.  During value coercion,
we map the string ids into integer ids, and then add the integer id for
the JSONDataModule if it is missing.