Makes the post_syncdb() functions idempotent
Previously, they were run-once functions.  This works fine when starting
a site or when running ./manage.py test with south migrations enabled;
in the latter case, ./manage.py test is overridden from the original
Django implementation, and runs ./manage.py syncdb, followed by an
overridden no-op ./manage.py flush, followed by ./manage.py migrate.

However, when south is not enabled during tests, the Django
implementation runs ./manage.py syncdb, followed by ./manage.py flush.
The ./manage.py flush command is also executed before execution of every
test method of a TransactionTestCase subclass.  The post_syncdb signal
is sent after successful completion of both commands, but under the
previous code, the post_syncdb() functions are not run after the
starting ./manage.py syncdb.  This means that at the start of testing,
and at the start of TransactionTestCase test methods, the database is
empty and missing its required data, which was supposed to be installed
by the post_syncdb() functions.

This commit makes the post_syncdb() functions idempotent, so that they
can be run successfully multiple times.  After a post_syncdb() function
is called (with the correct app as an argument), it is now guaranteed
that the initial data will be correctly installed in the database.

In the instances where this wasn't already the case, the install()
functions were also changed to be idempotent, in order to guarantee the
idempotence of the post_syncdb() functions.