Fix a minor bug in onsite class list
The class categories would get duplicated everytime the table was
rendered.  Previously, the class categories options were cleared each time
to avoid this problem.  That conflicted with the addition of "show all"
and "hide all" buttons to that area in b528d02 (which I'm guessing is why
the line top_div.html("") was commented out).  Moving the show/hide
buttons up one level in the DOM avoids the issue.