Adds missing uses of @functools.wraps()
In Django security release 1.4.11, reverse() stopped supporting paths
that weren't already loaded in a urlconf. The
RegexURLResolver._populate() method takes care of recursively inspecting
the root urlconf and maintaining a set _callback_strs of all paths used
in the urlconf that are allowed to be used in calls to reverse().

When a callable view function object is passed to the urlconf (as
opposed to a dotted path string), _populate() inspects the object to
determine the dotted path string from its module, class, and function
names. When the callable view function is decorated with a wrapper that
does not use @functools.wraps(), this information is lost. The result is
that legitimate calls to reverse() and {% url %} are no longer allowed.

This was causing an error in the admin panel. Specifically, the call to
{% url esp.themes.views.landing %}, since landing() is decorated by
@admin_required(), which was missing the use of @functools.wraps().

The fix is to make sure to always use @functools.wraps() when defining
decorators, especially when those decorators will be decorating view
functions.

Another possible fix would have been to use the wrapt package. See
Github issue #1307.