Allow morphed users to bypass required modules
This small change addresses #1097 (the latter option: "non-profile
required modules shouldn't take effect while onsite morphing).

Submitting as a separate pull request so someone can sanity check the
logic.  I'm not 100% sure this is a good idea; it means there is more
responsibility for the onsite volunteer to make sure the student is
authorized for whatever changes they are making (the Web site won't slow
them down).  It also means that the student has to walk the onsite
volunteer through the whole profile form if, for some reason, they have no
profile.