Modify accounting controller to consider line item options
The get_preferences(), set_preferences(), and apply_preferences() methods
have an informal I/O format of "tuples of human-readable information about
the transfers you have/want."  Along those lines, this format has been
extended by adding a 4th element to the tuple, which is the ID of a
LineItemOptions associated with the item (or None).

There is some added logic to properly determine the dollar amount
according to the transfer.  If a dollar amount is specified in the input,
that is used (overridding both the line item type, and any chosen line
item options).  If a LineItemOptions ID is provied, its dollar amount is
used (overriding the line item type).  Otherwise the dollar amount is
copied from the line item type.