Set transaction ID after charging payment
In 406781a we reordered things to save a record in the database (via the
accounting controller) before charging the credit card using
stripe.Charge.create().  I think this makes sense, but there was a slight
problem: the transaction ID (charge.id) didn't exist yet when calling
iac.submit_payment().

This is a small tweak to fix the error while preserving the general
workflow - iac.submit_payment() is called with a transaction ID of 'TBD'
but we get the Transfer object and save it again once we know the real
transaction ID.