Fixes CommunicationsPanelTest
As a result of the change in 20f4ec3c, some of the test emails were
being de-duplicated.  This was happening because
ProgramFrameworkTest.setUp() was creating test users with the same name
and email address (all empty strings, in fact).

This change has setUp() create test users with unique names.