Specify program when fetching/creating donation line item type
The lack of a program was causing an error if this function was creating
the line item type (instead of just fetching it).  The program and text
are the only required fields on LineItemType; we don't need to specify an
amount_dec since the user has to explicitly choose the amount.