Adds phone numbers back to teacher checkin page
833b201b introduced a regression on the teacher checkin page,
accidentally removing the phone numbers of missing teachers. This is
because phone numbers were being set by iterating over
ClassSection.teachers and setting a field on those ESPUser objects. This
doesn't work because `teachers` in this context is a property, so the
assignments of phone numbers were quickly lost.

This is fixed by creating a list of teachers and then modifying that. It
is this list which is now iterated over in the template.

This commit also:
- Adds test_phone_numbers_on_checkin_page() to catch this regression.
- To support this test, adds ProgramFrameworkTest.add_user_profiles() to
  create teacher profiles with phone numbers.
- Provides default text for when the teacher has no contact info, or for
  when the teacher has no provided phone number.