Add student registration "big board"
When student registrations are coming in, it's fun to watch how many students
have registered.  Unfortunately, 10 admins refreshing the dashboard is not the
greatest at making the site fast for users.  This page should be quick to load,
and is cached for a bit under 2 minutes, so many admins loading it is not a
problem.

There's a slight performance issue in one of the queries; I believe it
shouldn't be a problem in production, and it will be hard to fix until we are
on django 1.5.  See the comment starting on line 128 of
program/modules/handlers/bigboardmodule.py.  Other than that, the page is
fairly quick to load, and everything is cached.