Renames OverrideFull permission
With the "Student/" prefix, the permission would be granted recursively
when "Students/All" was granted. Since this is a special override
permission, this behavior should not be possible. Renaming it to just
"OverrideFull" prevents this unwanted behavior. Now the permission is
only granted to admins, and to users who have specifically been given
the permission.

Also creates Program.OVERRIDE_FULL_PERM to hold this string constant.