Refactors permissions checks
It is now possible to query for the set of users that have a permission
on a program, with the Permission.users_with_perm() classmethod.

To reduce code duplication, Permission.user_has_perm() and
ESPUser.isAdministrator() now call this function.

Adds more checks in PermissionTestCase, to make sure that
ESPUser.isAdministrator() still works as expected.

This implementation might not be performant. Someone should check that
this doesn't degrade performance of Permissions.