Modifies the calculation of Program.isFull()
This calculation now ignores any students with the OverrideFull
permission. So if the cap gets hit, but some of those students have the
permission, then students without the permission can continue
registering until the number of students without the permission hits the
capacity.

Also modifies ESPUser.canRegToFullProgram(), so that even when a program
has reached capacity, the students who are already registered can
continue to access the program, modify their registrations, etc.