Changes program size count to be based on profiles
The old count was based on StudentRegistrations. This is slightly better
than the old count, because it means that if you've gotten to the screen
where you can select classes, you're guaranteed to be able to continue
selecting classes. Also, StudentRegistrations aren't a perfect indicator
of registration anymore, since there are now also
StudentSubjectInterests.

However, this is still non-ideal. In the past, we've noticed that for
some reason, the number of completed profiles is much greater than the
number of registrations. Additionally, if a student drops out and
deletes their registrations, their profile will still exist, so they
won't free up a spot. Both of these factors means that the program may
artificially appear full, even though the program could take more
students.

Despite these downsides, I'm not sure there's a better solution at the
moment.