Refactor the thunk-handling code in ArgCache
This is a first step toward removing the cache_loader module. The
planned replacement is something similar to the way Django handles
related field definitions; see add_lazy_relation in
django.db.models.fields.related.

This commit adds a new method add_lazy_dependency, which (similarly to
add_lazy_relation) encapsulates the pattern of delaying resolving
dependencies on models until those models are imported. Previously
this was done by the combination of handle_thunk and delay_method. The
eventual plan is for add_lazy_dependency to replace them both, but for
now, add_lazy_dependency borrows the thunking scheme of handle_thunk
and calls into delay_method. This is awkward, but is done for the sake
of changing less code at once.