Student lunch selection form now displays already chosen lunch blocks correctly.
Fixes #1453.
Note: in the lottery, if students fill out this form while the lottery is open, it will let them submit multiple lunch blocks, and won't expire old ones. This might lead to an unexpected display of the form (right now it just gets the first one retrieved in the query for that day).