Fix cache dependencies of render_class
We need to depend on not only the user's StudentRegistrations for this
section, but in fact on their StudentRegistrations for all sections, because
of things like lunch constraints -- a change made in another block could
affect whether you can add a class in this one.  So we depend on all SRs for
this user.

This was causing some nasty caching bugs in lunch constraints for MIT.