Adding migration for CRMI cleanup
South wanted a bunch more things to be in this migration. One of
them was apply_multiplier_to_room_cap, which I remember adding a year
ago and there's already a migration for it. Another one was something
about the SATPrepTeacherModuleInfo, which I remember running into
when I was trying to make the migration for apply_multiplier_to_room_cap
a year ago. I don't know why it's doing this, but the migration
I'm committing here ran fine, so it's probably fine...until the next time
someone tries to change the CRMI or SCRMI and gets confused.