Simplify validation of ChoiceWithOtherField
This field was bravely attempting to validate whether a custom dollar
amount had been entered for line item options, if the last option
supported custom amounts.  This is a complicated issue since it's possible
for multiple options to support custom amounts.  So I'm removing that
check.

If no custom amount is specified, the user's saved preferences will
include the default dollar amount for the line item option or line item
type.  (The safe bet for program directors is to make this 0.)

Also, fixes a bug where the ordering of the output tuple was reversed.