Remove old email verify module
It had no @main_call or @aux_call, nor even a method with the right signature;
the method that appeared to be intended as its @main_call contained undefined
names, it didn't even exist as a ProgramModule in the esp.mit.edu database.
The code does not appear to have been nontrivially touched since at least 2007.