Add fields to select_related in getLastForProgram
Not 100% sure that this is a good idea, but I would expect that since we're
only ever returning a single RegistrationProfile the .select_related() will be
fast even though it hits a bunch more tables, since it really just needs to
do a single indexed lookup in each.  This will save a query per user in the
dashboard stats, and may save queries elsewhere too; I bet this isn't the only
place get a RegistrationProfile and then immediately grab one of its infos.