Make teacherreg do more sane things with default resource types
The program creation form already creates the default resource
types. If someone deleted one, it's probably because they didn't
want it. The logic for the non-static case is still kind of silly
but I didn't want to do much with it because I don't know what
the desired behavior there is.