Reset 'register student' dialog to proper state
If, for example, you were trying to migrate a student into the program,
and they were missing some required profile information, the dialog would
correctly tell you to use the other onsite interface.  But if you tried to
repeat this operation for other students, the dialog would open up in the
"error message" state.  This is now fixed, and also associates specific
IDs with the buttons so the "register" button is disabled in the error
state.