Use jQueryUI's autocompletechange instead of the native change event.
This fixes a bug in the AjaxForeignKeyNewformField. Previously, the
underlying value was cleared whenever the change event was fired on
the textbox, the intent being that this should happen when users
manually clear or edit the textbox. However, in some circumstances,
the change event would be fired after selecting an option from the
autocomplete, causing the just-selected value to be clobbered. (This
would sometimes happen in Firefox depending on what the user did after
selecting the option.)

Instead of clearing the value on the native change event, use the
autocompletechange event. This has two effects. First, it is more
supported by jQuery UI, and therefore might behave more consistently
across browsers. Second, the autocompletechange handler has access to
the currently-selected option through ui.item, so rather than clearing
the value all the time, we can clear it only when ui.item is null
(i.e. no item is selected), which is what we want.