Remove uses of async: false
This prevents the behavior where adding/removing students freezes the
browser until the server responds with the result.

Assuming this was originally written this way in order to avoid
reasoning about multiple requests to the server happening at the same
time and possibly getting mixed up, add some locking to disable the
checkboxes while a request to add/remove a student is outstanding.

Fixes #1547.