Update `ClassSection.viable_rooms()`
It now uses locations, and is a bit smarter.  I removed caching, both because
the function should be less terribly slow, and because it would need a lot of
dependencies (in particular, the scheduling of any class anywhere), and it's
not clear that's worth it, especially since it's currently only called in
tests.