Merge the old and new lottery_preferences formats
In the case of a single priority, lottery_preferences used to return
data in a different format, but the two-phase lottery expected the
same format as for multiple priorities. This change makes it
compatible with functions that expect either format by returning a
merge of the two formats.

Also add a comment explaining that the old format is deprecated and
should not be used.