Adds useful script to run a query against each chapter site
I was tired of having questions like "how many chapters have X object in their
database" and having no easy way to answer them.  This script is janky but
should hopefully work; maybe some day we will have an actual build system that
can do this reasonably.  Note that while the script lives in the codebase of
one site, it is designed to run on all sites, I just don't have a better place
to keep it.