Fix getResourceTypes and getResources
This is a little sneaky: we will now only have behavior as if
include_classrooms is False in getResourceTypes.  The reason this is safe, even
though it was sometimes previously called with include_classrooms=True, is that
Classroom must always be a global resource (due to the number of calls of the
form `ResourceType.get_or_create('Classroom')`), and allow_global_restypes is
False on all sites.  So right now, the behavior on all sites is to exclude
classrooms.  So we can just make that always the case, even if global resource
types later become allowed, and no one will care.

getResources works the same way, except it always excludes global resource
types already.