Change add_lazy_dependency to use apps.get_registered_model
This is again modeled after Django's add_lazy_relation (see
62d88c9aedfd8be44ec51c79feba935092c16810) as implemented in 1.7+. The
docstring of get_registered_model claims that it is similar to
get_model, but safe to call while the registry is still being
populated.

This is still a private API, though, and it's probably better to
eventually switch to using the AppConfig.ready() method to set up
the cache dependencies.