Pipe input directly to lessc to avoid making a temp file.
I vaguely remember people in the past encountering some sort of
permissions issue involving temporary files during themes setup. I
don't know if it's still prone to happen or if this is even the
relevant code, but this might address that issue and seems to
otherwise work the same as before.