Remove use of subprocess.PIPE in gen_latex()
`stdout=PIPE` was used as a shortcut to hide the output of latex
generation. However, according to the documentation for
`subprocess.call()` [1], this is unsafe, as "that can deadlock
based on the child process output volume".

This explains some of the symptomps of #1198. Namely, a hanging
pdflatex process, for what should normally be a <10 second
operation.

Instead, the default arguments to `gen_latex()` now open a file
handle to `os.devnull`, and pass that to `subprocess.call()`.

Fixes, either partially or entirely, #1198 (depending on whether
database deadlocks from accounting are still an issue).

[1] <https://docs.python.org/2/library/subprocess.html#subprocess.call>